did the requested changes (made the checkpointsession table not nullable) as well as fixed some stuff regarding the mappers and changing Ids from Int to Long

+++ src/main/kotlin/com/scrumdapp/checkpointservice/dto/CheckpointDto.kt
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.annotation.JsonInclude
 @JsonInclude(JsonInclude.Include.NON_NULL)
 data class CheckpointResponseDto(
     val id: Long,
-    val sessionId: Int,
-    val groupUserId: Int,
+    val sessionId: Long,
+    val groupUserId: Long,
     val presence: Int?,
     val impediment: String?,
     val stars: Int?,
